Rename calculation functions
Include 'Trade' in the name
