+++ modules/measure/src/main/java/com/opengamma/strata/measure/deposit/TermDepositTradeCalculationFunction.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.payment;
+package com.opengamma.strata.measure.deposit;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -25,12 +25,12 @@ import com.opengamma.strata.data.scenario.ScenarioMarketData;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.measure.rate.RatesScenarioMarketData;
-import com.opengamma.strata.product.payment.BulletPayment;
-import com.opengamma.strata.product.payment.BulletPaymentTrade;
-import com.opengamma.strata.product.payment.ResolvedBulletPaymentTrade;
+import com.opengamma.strata.product.deposit.ResolvedTermDepositTrade;
+import com.opengamma.strata.product.deposit.TermDeposit;
+import com.opengamma.strata.product.deposit.TermDepositTrade;
 
 /**
- * Perform calculations on a single {@code BulletPaymentTrade} for each of a set of scenarios.
+ * Perform calculations on a single {@code TermDepositTrade} for each of a set of scenarios.
  * <p>
  * This uses the standard discounting calculation method.
  * The supported built-in measures are:
@@ -43,17 +43,19 @@ import com.opengamma.strata.product.payment.ResolvedBulletPaymentTrade;
  *   <li>{@linkplain Measures#BUCKETED_PV01 Bucketed PV01}
  * </ul>
  */
-public class BulletPaymentCalculationFunction
-    implements CalculationFunction<BulletPaymentTrade> {
+public class TermDepositTradeCalculationFunction
+    implements CalculationFunction<TermDepositTrade> {
 
   /**
    * The calculations by measure.
    */
   private static final ImmutableMap<Measure, SingleMeasureCalculation> CALCULATORS =
       ImmutableMap.<Measure, SingleMeasureCalculation>builder()
-          .put(Measures.PRESENT_VALUE, BulletPaymentMeasureCalculations::presentValue)
-          .put(Measures.PV01, BulletPaymentMeasureCalculations::pv01)
-          .put(Measures.BUCKETED_PV01, BulletPaymentMeasureCalculations::bucketedPv01)
+          .put(Measures.PAR_RATE, TermDepositMeasureCalculations::parRate)
+          .put(Measures.PAR_SPREAD, TermDepositMeasureCalculations::parSpread)
+          .put(Measures.PRESENT_VALUE, TermDepositMeasureCalculations::presentValue)
+          .put(Measures.PV01, TermDepositMeasureCalculations::pv01)
+          .put(Measures.BUCKETED_PV01, TermDepositMeasureCalculations::bucketedPv01)
           .build();
 
   private static final ImmutableSet<Measure> MEASURES = ImmutableSet.<Measure>builder()
@@ -64,13 +66,13 @@ public class BulletPaymentCalculationFunction
   /**
    * Creates an instance.
    */
-  public BulletPaymentCalculationFunction() {
+  public TermDepositTradeCalculationFunction() {
   }
 
   //-------------------------------------------------------------------------
   @Override
-  public Class<BulletPaymentTrade> targetType() {
-    return BulletPaymentTrade.class;
+  public Class<TermDepositTrade> targetType() {
+    return TermDepositTrade.class;
   }
 
   @Override
@@ -79,20 +81,20 @@ public class BulletPaymentCalculationFunction
   }
 
   @Override
-  public Currency naturalCurrency(BulletPaymentTrade trade, ReferenceData refData) {
+  public Currency naturalCurrency(TermDepositTrade trade, ReferenceData refData) {
     return trade.getProduct().getCurrency();
   }
 
   //-------------------------------------------------------------------------
   @Override
   public FunctionRequirements requirements(
-      BulletPaymentTrade trade,
+      TermDepositTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ReferenceData refData) {
 
     // extract data from product
-    BulletPayment product = trade.getProduct();
+    TermDeposit product = trade.getProduct();
     Currency currency = product.getCurrency();
 
     // use lookup to build requirements
@@ -103,14 +105,14 @@ public class BulletPaymentCalculationFunction
   //-------------------------------------------------------------------------
   @Override
   public Map<Measure, Result<?>> calculate(
-      BulletPaymentTrade trade,
+      TermDepositTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ScenarioMarketData scenarioMarketData,
       ReferenceData refData) {
 
     // resolve the trade once for all measures and all scenarios
-    ResolvedBulletPaymentTrade resolved = trade.resolve(refData);
+    ResolvedTermDepositTrade resolved = trade.resolve(refData);
 
     // use lookup to query market data
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
@@ -129,7 +131,7 @@ public class BulletPaymentCalculationFunction
   // calculate one measure
   private Result<?> calculate(
       Measure measure,
-      ResolvedBulletPaymentTrade trade,
+      ResolvedTermDepositTrade trade,
       RatesScenarioMarketData marketData) {
 
     SingleMeasureCalculation calculator = CALCULATORS.get(measure);
@@ -143,7 +145,7 @@ public class BulletPaymentCalculationFunction
   @FunctionalInterface
   interface SingleMeasureCalculation {
     public abstract ScenarioArray<?> calculate(
-        ResolvedBulletPaymentTrade trade,
+        ResolvedTermDepositTrade trade,
         RatesScenarioMarketData marketData);
   }
