+++ modules/measure/src/main/java/com/opengamma/strata/measure/fx/FxSingleTradeCalculationFunction.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.index;
+package com.opengamma.strata.measure.fx;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -13,7 +13,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.opengamma.strata.basics.ReferenceData;
 import com.opengamma.strata.basics.currency.Currency;
-import com.opengamma.strata.basics.index.IborIndex;
+import com.opengamma.strata.basics.currency.CurrencyPair;
 import com.opengamma.strata.calc.Measure;
 import com.opengamma.strata.calc.runner.CalculationFunction;
 import com.opengamma.strata.calc.runner.CalculationParameters;
@@ -21,20 +21,17 @@ import com.opengamma.strata.calc.runner.FunctionRequirements;
 import com.opengamma.strata.calc.runner.FunctionUtils;
 import com.opengamma.strata.collect.result.FailureReason;
 import com.opengamma.strata.collect.result.Result;
-import com.opengamma.strata.data.FieldName;
-import com.opengamma.strata.data.MarketDataId;
 import com.opengamma.strata.data.scenario.ScenarioArray;
 import com.opengamma.strata.data.scenario.ScenarioMarketData;
-import com.opengamma.strata.market.observable.QuoteId;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.measure.rate.RatesScenarioMarketData;
-import com.opengamma.strata.product.index.IborFuture;
-import com.opengamma.strata.product.index.IborFutureTrade;
-import com.opengamma.strata.product.index.ResolvedIborFutureTrade;
+import com.opengamma.strata.product.fx.FxSingle;
+import com.opengamma.strata.product.fx.FxSingleTrade;
+import com.opengamma.strata.product.fx.ResolvedFxSingleTrade;
 
 /**
- * Perform calculations on a single {@code IborFutureTrade} for each of a set of scenarios.
+ * Perform calculations on a single {@code FxSingleTrade} for each of a set of scenarios.
  * <p>
  * This uses the standard discounting calculation method.
  * The supported built-in measures are:
@@ -44,20 +41,28 @@ import com.opengamma.strata.product.index.ResolvedIborFutureTrade;
  *   <li>{@linkplain Measures#PRESENT_VALUE_MULTI_CCY Present value with no currency conversion}
  *   <li>{@linkplain Measures#PV01 PV01}
  *   <li>{@linkplain Measures#BUCKETED_PV01 Bucketed PV01}
+ *   <li>{@linkplain Measures#CURRENCY_EXPOSURE Currency exposure}
+ *   <li>{@linkplain Measures#CURRENT_CASH Current cash}
+ *   <li>{@linkplain Measures#FORWARD_FX_RATE Forward FX rate}
  * </ul>
+ * <p>
+ * The "natural" currency is the base currency of the market convention pair of the two trade currencies.
  */
-public class IborFutureCalculationFunction
-    implements CalculationFunction<IborFutureTrade> {
+public class FxSingleTradeCalculationFunction
+    implements CalculationFunction<FxSingleTrade> {
 
   /**
    * The calculations by measure.
    */
   private static final ImmutableMap<Measure, SingleMeasureCalculation> CALCULATORS =
       ImmutableMap.<Measure, SingleMeasureCalculation>builder()
-          .put(Measures.PAR_SPREAD, IborFutureMeasureCalculations::parSpread)
-          .put(Measures.PRESENT_VALUE, IborFutureMeasureCalculations::presentValue)
-          .put(Measures.PV01, IborFutureMeasureCalculations::pv01)
-          .put(Measures.BUCKETED_PV01, IborFutureMeasureCalculations::bucketedPv01)
+          .put(Measures.PAR_SPREAD, FxSingleMeasureCalculations::parSpread)
+          .put(Measures.PRESENT_VALUE, FxSingleMeasureCalculations::presentValue)
+          .put(Measures.PV01, FxSingleMeasureCalculations::pv01)
+          .put(Measures.BUCKETED_PV01, FxSingleMeasureCalculations::bucketedPv01)
+          .put(Measures.CURRENCY_EXPOSURE, FxSingleMeasureCalculations::currencyExposure)
+          .put(Measures.CURRENT_CASH, FxSingleMeasureCalculations::currentCash)
+          .put(Measures.FORWARD_FX_RATE, FxSingleMeasureCalculations::forwardFxRate)
           .build();
 
   private static final ImmutableSet<Measure> MEASURES = ImmutableSet.<Measure>builder()
@@ -68,13 +73,13 @@ public class IborFutureCalculationFunction
   /**
    * Creates an instance.
    */
-  public IborFutureCalculationFunction() {
+  public FxSingleTradeCalculationFunction() {
   }
 
   //-------------------------------------------------------------------------
   @Override
-  public Class<IborFutureTrade> targetType() {
-    return IborFutureTrade.class;
+  public Class<FxSingleTrade> targetType() {
+    return FxSingleTrade.class;
   }
 
   @Override
@@ -83,45 +88,43 @@ public class IborFutureCalculationFunction
   }
 
   @Override
-  public Currency naturalCurrency(IborFutureTrade trade, ReferenceData refData) {
-    return trade.getProduct().getCurrency();
+  public Currency naturalCurrency(FxSingleTrade trade, ReferenceData refData) {
+    Currency base = trade.getProduct().getBaseCurrencyAmount().getCurrency();
+    Currency counter = trade.getProduct().getCounterCurrencyAmount().getCurrency();
+    CurrencyPair marketConventionPair = CurrencyPair.of(base, counter).toConventional();
+    return marketConventionPair.getBase();
   }
 
   //-------------------------------------------------------------------------
   @Override
   public FunctionRequirements requirements(
-      IborFutureTrade trade,
+      FxSingleTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ReferenceData refData) {
 
     // extract data from product
-    IborFuture product = trade.getProduct();
-    QuoteId quoteId = QuoteId.of(trade.getProduct().getSecurityId().getStandardId(), FieldName.SETTLEMENT_PRICE);
-    Currency currency = product.getCurrency();
-    IborIndex index = product.getIndex();
+    FxSingle fx = trade.getProduct();
+    Currency baseCurrency = fx.getBaseCurrencyAmount().getCurrency();
+    Currency counterCurrency = fx.getCounterCurrencyAmount().getCurrency();
+    ImmutableSet<Currency> currencies = ImmutableSet.of(baseCurrency, counterCurrency);
 
     // use lookup to build requirements
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
-    FunctionRequirements ratesReqs = ratesLookup.requirements(currency, index);
-    ImmutableSet<MarketDataId<?>> valueReqs = ImmutableSet.<MarketDataId<?>>builder()
-        .add(quoteId)
-        .addAll(ratesReqs.getValueRequirements())
-        .build();
-    return ratesReqs.toBuilder().valueRequirements(valueReqs).build();
+    return ratesLookup.requirements(currencies);
   }
 
   //-------------------------------------------------------------------------
   @Override
   public Map<Measure, Result<?>> calculate(
-      IborFutureTrade trade,
+      FxSingleTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ScenarioMarketData scenarioMarketData,
       ReferenceData refData) {
 
     // resolve the trade once for all measures and all scenarios
-    ResolvedIborFutureTrade resolved = trade.resolve(refData);
+    ResolvedFxSingleTrade resolved = trade.resolve(refData);
 
     // use lookup to query market data
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
@@ -140,7 +143,7 @@ public class IborFutureCalculationFunction
   // calculate one measure
   private Result<?> calculate(
       Measure measure,
-      ResolvedIborFutureTrade trade,
+      ResolvedFxSingleTrade trade,
       RatesScenarioMarketData marketData) {
 
     SingleMeasureCalculation calculator = CALCULATORS.get(measure);
@@ -154,7 +157,7 @@ public class IborFutureCalculationFunction
   @FunctionalInterface
   interface SingleMeasureCalculation {
     public abstract ScenarioArray<?> calculate(
-        ResolvedIborFutureTrade trade,
+        ResolvedFxSingleTrade trade,
         RatesScenarioMarketData marketData);
   }
