+++ modules/measure/src/main/java/com/opengamma/strata/measure/fx/FxSwapTradeCalculationFunction.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.dsf;
+package com.opengamma.strata.measure.fx;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -13,7 +13,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.opengamma.strata.basics.ReferenceData;
 import com.opengamma.strata.basics.currency.Currency;
-import com.opengamma.strata.basics.index.Index;
+import com.opengamma.strata.basics.currency.CurrencyPair;
 import com.opengamma.strata.calc.Measure;
 import com.opengamma.strata.calc.runner.CalculationFunction;
 import com.opengamma.strata.calc.runner.CalculationParameters;
@@ -21,43 +21,47 @@ import com.opengamma.strata.calc.runner.FunctionRequirements;
 import com.opengamma.strata.calc.runner.FunctionUtils;
 import com.opengamma.strata.collect.result.FailureReason;
 import com.opengamma.strata.collect.result.Result;
-import com.opengamma.strata.data.FieldName;
-import com.opengamma.strata.data.MarketDataId;
 import com.opengamma.strata.data.scenario.ScenarioArray;
 import com.opengamma.strata.data.scenario.ScenarioMarketData;
-import com.opengamma.strata.market.observable.QuoteId;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.measure.rate.RatesScenarioMarketData;
-import com.opengamma.strata.product.dsf.Dsf;
-import com.opengamma.strata.product.dsf.DsfTrade;
-import com.opengamma.strata.product.dsf.ResolvedDsfTrade;
+import com.opengamma.strata.product.fx.FxSwap;
+import com.opengamma.strata.product.fx.FxSwapTrade;
+import com.opengamma.strata.product.fx.ResolvedFxSwapTrade;
 
 /**
- * Perform calculations on a single {@code DsfTrade} for each of a set of scenarios.
+ * Perform calculations on a single {@code FxSwapTrade} for each of a set of scenarios.
  * <p>
  * This uses the standard discounting calculation method.
  * The supported built-in measures are:
  * <ul>
+ *   <li>{@linkplain Measures#PAR_SPREAD Par spread}
  *   <li>{@linkplain Measures#PRESENT_VALUE Present value}
  *   <li>{@linkplain Measures#PRESENT_VALUE_MULTI_CCY Present value with no currency conversion}
  *   <li>{@linkplain Measures#PV01 PV01}
  *   <li>{@linkplain Measures#BUCKETED_PV01 Bucketed PV01}
+ *   <li>{@linkplain Measures#CURRENCY_EXPOSURE Currency exposure}
+ *   <li>{@linkplain Measures#CURRENT_CASH Current cash}
+ *   <li>{@linkplain Measures#FORWARD_FX_RATE Forward FX rate}
  * </ul>
  * <p>
- * The "natural" currency is the currency of the swap leg that is received.
+ * The "natural" currency is the base currency of the market convention pair of the near leg currencies.
  */
-public class DsfCalculationFunction
-    implements CalculationFunction<DsfTrade> {
+public class FxSwapTradeCalculationFunction
+    implements CalculationFunction<FxSwapTrade> {
 
   /**
    * The calculations by measure.
    */
   private static final ImmutableMap<Measure, SingleMeasureCalculation> CALCULATORS =
       ImmutableMap.<Measure, SingleMeasureCalculation>builder()
-          .put(Measures.PRESENT_VALUE, DsfMeasureCalculations::presentValue)
-          .put(Measures.PV01, DsfMeasureCalculations::pv01)
-          .put(Measures.BUCKETED_PV01, DsfMeasureCalculations::bucketedPv01)
+          .put(Measures.PAR_SPREAD, FxSwapMeasureCalculations::parSpread)
+          .put(Measures.PRESENT_VALUE, FxSwapMeasureCalculations::presentValue)
+          .put(Measures.PV01, FxSwapMeasureCalculations::pv01)
+          .put(Measures.BUCKETED_PV01, FxSwapMeasureCalculations::bucketedPv01)
+          .put(Measures.CURRENCY_EXPOSURE, FxSwapMeasureCalculations::currencyExposure)
+          .put(Measures.CURRENT_CASH, FxSwapMeasureCalculations::currentCash)
           .build();
 
   private static final ImmutableSet<Measure> MEASURES = ImmutableSet.<Measure>builder()
@@ -68,13 +72,13 @@ public class DsfCalculationFunction
   /**
    * Creates an instance.
    */
-  public DsfCalculationFunction() {
+  public FxSwapTradeCalculationFunction() {
   }
 
   //-------------------------------------------------------------------------
   @Override
-  public Class<DsfTrade> targetType() {
-    return DsfTrade.class;
+  public Class<FxSwapTrade> targetType() {
+    return FxSwapTrade.class;
   }
 
   @Override
@@ -83,45 +87,43 @@ public class DsfCalculationFunction
   }
 
   @Override
-  public Currency naturalCurrency(DsfTrade trade, ReferenceData refData) {
-    return trade.getProduct().getCurrency();
+  public Currency naturalCurrency(FxSwapTrade trade, ReferenceData refData) {
+    Currency base = trade.getProduct().getNearLeg().getBaseCurrencyAmount().getCurrency();
+    Currency counter = trade.getProduct().getNearLeg().getCounterCurrencyAmount().getCurrency();
+    CurrencyPair marketConventionPair = CurrencyPair.of(base, counter).toConventional();
+    return marketConventionPair.getBase();
   }
 
   //-------------------------------------------------------------------------
   @Override
   public FunctionRequirements requirements(
-      DsfTrade trade,
+      FxSwapTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ReferenceData refData) {
 
     // extract data from product
-    Dsf product = trade.getProduct();
-    QuoteId quoteId = QuoteId.of(trade.getSecurityId().getStandardId(), FieldName.SETTLEMENT_PRICE);
-    ImmutableSet<Index> indices = product.getUnderlyingSwap().allIndices();
-    ImmutableSet<Currency> currencies = ImmutableSet.of(product.getCurrency());
+    FxSwap fx = trade.getProduct();
+    Currency baseCurrency = fx.getNearLeg().getBaseCurrencyAmount().getCurrency();
+    Currency counterCurrency = fx.getNearLeg().getCounterCurrencyAmount().getCurrency();
+    ImmutableSet<Currency> currencies = ImmutableSet.of(baseCurrency, counterCurrency);
 
     // use lookup to build requirements
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
-    FunctionRequirements ratesReqs = ratesLookup.requirements(currencies, indices);
-    ImmutableSet<MarketDataId<?>> valueReqs = ImmutableSet.<MarketDataId<?>>builder()
-        .add(quoteId)
-        .addAll(ratesReqs.getValueRequirements())
-        .build();
-    return ratesReqs.toBuilder().valueRequirements(valueReqs).build();
+    return ratesLookup.requirements(currencies);
   }
 
   //-------------------------------------------------------------------------
   @Override
   public Map<Measure, Result<?>> calculate(
-      DsfTrade trade,
+      FxSwapTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ScenarioMarketData scenarioMarketData,
       ReferenceData refData) {
 
     // resolve the trade once for all measures and all scenarios
-    ResolvedDsfTrade resolved = trade.resolve(refData);
+    ResolvedFxSwapTrade resolved = trade.resolve(refData);
 
     // use lookup to query market data
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
@@ -140,7 +142,7 @@ public class DsfCalculationFunction
   // calculate one measure
   private Result<?> calculate(
       Measure measure,
-      ResolvedDsfTrade trade,
+      ResolvedFxSwapTrade trade,
       RatesScenarioMarketData marketData) {
 
     SingleMeasureCalculation calculator = CALCULATORS.get(measure);
@@ -154,7 +156,7 @@ public class DsfCalculationFunction
   @FunctionalInterface
   interface SingleMeasureCalculation {
     public abstract ScenarioArray<?> calculate(
-        ResolvedDsfTrade trade,
+        ResolvedFxSwapTrade trade,
         RatesScenarioMarketData marketData);
   }
