+++ modules/measure/src/main/java/com/opengamma/strata/measure/index/IborFutureTradeCalculationFunction.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.swap;
+package com.opengamma.strata.measure.index;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -13,6 +13,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.opengamma.strata.basics.ReferenceData;
 import com.opengamma.strata.basics.currency.Currency;
+import com.opengamma.strata.basics.index.IborIndex;
 import com.opengamma.strata.calc.Measure;
 import com.opengamma.strata.calc.runner.CalculationFunction;
 import com.opengamma.strata.calc.runner.CalculationParameters;
@@ -20,60 +21,43 @@ import com.opengamma.strata.calc.runner.FunctionRequirements;
 import com.opengamma.strata.calc.runner.FunctionUtils;
 import com.opengamma.strata.collect.result.FailureReason;
 import com.opengamma.strata.collect.result.Result;
+import com.opengamma.strata.data.FieldName;
+import com.opengamma.strata.data.MarketDataId;
 import com.opengamma.strata.data.scenario.ScenarioArray;
 import com.opengamma.strata.data.scenario.ScenarioMarketData;
+import com.opengamma.strata.market.observable.QuoteId;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.measure.rate.RatesScenarioMarketData;
-import com.opengamma.strata.product.swap.ResolvedSwapTrade;
-import com.opengamma.strata.product.swap.Swap;
-import com.opengamma.strata.product.swap.SwapTrade;
+import com.opengamma.strata.product.index.IborFuture;
+import com.opengamma.strata.product.index.IborFutureTrade;
+import com.opengamma.strata.product.index.ResolvedIborFutureTrade;
 
 /**
- * Perform calculations on a single {@code SwapTrade} for each of a set of scenarios.
+ * Perform calculations on a single {@code IborFutureTrade} for each of a set of scenarios.
  * <p>
  * This uses the standard discounting calculation method.
  * The supported built-in measures are:
  * <ul>
- *   <li>{@linkplain Measures#PAR_RATE Par rate}
  *   <li>{@linkplain Measures#PAR_SPREAD Par spread}
  *   <li>{@linkplain Measures#PRESENT_VALUE Present value}
  *   <li>{@linkplain Measures#PRESENT_VALUE_MULTI_CCY Present value with no currency conversion}
- *   <li>{@linkplain Measures#EXPLAIN_PRESENT_VALUE Explain present value}
- *   <li>{@linkplain Measures#CASH_FLOWS Cash flows}
  *   <li>{@linkplain Measures#PV01 PV01}
  *   <li>{@linkplain Measures#BUCKETED_PV01 Bucketed PV01}
- *   <li>{@linkplain Measures#BUCKETED_GAMMA_PV01 Gamma PV01}
- *   <li>{@linkplain Measures#ACCRUED_INTEREST Accrued interest}
- *   <li>{@linkplain Measures#LEG_INITIAL_NOTIONAL Leg initial notional}
- *   <li>{@linkplain Measures#LEG_PRESENT_VALUE Leg present value}
- *   <li>{@linkplain Measures#CURRENCY_EXPOSURE Currency exposure}
- *   <li>{@linkplain Measures#CURRENT_CASH Current cash}
  * </ul>
- * <p>
- * The "natural" currency is the currency of the swaption, which is limited to be single-currency.
  */
-public class SwapCalculationFunction
-    implements CalculationFunction<SwapTrade> {
+public class IborFutureTradeCalculationFunction
+    implements CalculationFunction<IborFutureTrade> {
 
   /**
    * The calculations by measure.
    */
   private static final ImmutableMap<Measure, SingleMeasureCalculation> CALCULATORS =
       ImmutableMap.<Measure, SingleMeasureCalculation>builder()
-          .put(Measures.PAR_RATE, SwapMeasureCalculations::parRate)
-          .put(Measures.PAR_SPREAD, SwapMeasureCalculations::parSpread)
-          .put(Measures.PRESENT_VALUE, SwapMeasureCalculations::presentValue)
-          .put(Measures.EXPLAIN_PRESENT_VALUE, SwapMeasureCalculations::explainPresentValue)
-          .put(Measures.CASH_FLOWS, SwapMeasureCalculations::cashFlows)
-          .put(Measures.PV01, SwapMeasureCalculations::pv01)
-          .put(Measures.BUCKETED_PV01, SwapMeasureCalculations::bucketedPv01)
-          .put(Measures.BUCKETED_GAMMA_PV01, SwapMeasureCalculations::bucketedGammaPv01)
-          .put(Measures.ACCRUED_INTEREST, SwapMeasureCalculations::accruedInterest)
-          .put(Measures.LEG_INITIAL_NOTIONAL, SwapMeasureCalculations::legInitialNotional)
-          .put(Measures.LEG_PRESENT_VALUE, SwapMeasureCalculations::legPresentValue)
-          .put(Measures.CURRENCY_EXPOSURE, SwapMeasureCalculations::currencyExposure)
-          .put(Measures.CURRENT_CASH, SwapMeasureCalculations::currentCash)
+          .put(Measures.PAR_SPREAD, IborFutureMeasureCalculations::parSpread)
+          .put(Measures.PRESENT_VALUE, IborFutureMeasureCalculations::presentValue)
+          .put(Measures.PV01, IborFutureMeasureCalculations::pv01)
+          .put(Measures.BUCKETED_PV01, IborFutureMeasureCalculations::bucketedPv01)
           .build();
 
   private static final ImmutableSet<Measure> MEASURES = ImmutableSet.<Measure>builder()
@@ -84,13 +68,13 @@ public class SwapCalculationFunction
   /**
    * Creates an instance.
    */
-  public SwapCalculationFunction() {
+  public IborFutureTradeCalculationFunction() {
   }
 
   //-------------------------------------------------------------------------
   @Override
-  public Class<SwapTrade> targetType() {
-    return SwapTrade.class;
+  public Class<IborFutureTrade> targetType() {
+    return IborFutureTrade.class;
   }
 
   @Override
@@ -99,38 +83,45 @@ public class SwapCalculationFunction
   }
 
   @Override
-  public Currency naturalCurrency(SwapTrade trade, ReferenceData refData) {
-    return trade.getProduct().getLegs().get(0).getCurrency();
+  public Currency naturalCurrency(IborFutureTrade trade, ReferenceData refData) {
+    return trade.getProduct().getCurrency();
   }
 
   //-------------------------------------------------------------------------
   @Override
   public FunctionRequirements requirements(
-      SwapTrade trade,
+      IborFutureTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ReferenceData refData) {
 
     // extract data from product
-    Swap product = trade.getProduct();
-    ImmutableSet<Currency> currencies = product.allPaymentCurrencies();
+    IborFuture product = trade.getProduct();
+    QuoteId quoteId = QuoteId.of(trade.getProduct().getSecurityId().getStandardId(), FieldName.SETTLEMENT_PRICE);
+    Currency currency = product.getCurrency();
+    IborIndex index = product.getIndex();
 
     // use lookup to build requirements
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
-    return ratesLookup.requirements(currencies, product.allIndices());
+    FunctionRequirements ratesReqs = ratesLookup.requirements(currency, index);
+    ImmutableSet<MarketDataId<?>> valueReqs = ImmutableSet.<MarketDataId<?>>builder()
+        .add(quoteId)
+        .addAll(ratesReqs.getValueRequirements())
+        .build();
+    return ratesReqs.toBuilder().valueRequirements(valueReqs).build();
   }
 
   //-------------------------------------------------------------------------
   @Override
   public Map<Measure, Result<?>> calculate(
-      SwapTrade trade,
+      IborFutureTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ScenarioMarketData scenarioMarketData,
       ReferenceData refData) {
 
     // resolve the trade once for all measures and all scenarios
-    ResolvedSwapTrade resolved = trade.resolve(refData);
+    ResolvedIborFutureTrade resolved = trade.resolve(refData);
 
     // use lookup to query market data
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
@@ -149,7 +140,7 @@ public class SwapCalculationFunction
   // calculate one measure
   private Result<?> calculate(
       Measure measure,
-      ResolvedSwapTrade trade,
+      ResolvedIborFutureTrade trade,
       RatesScenarioMarketData marketData) {
 
     SingleMeasureCalculation calculator = CALCULATORS.get(measure);
@@ -163,7 +154,7 @@ public class SwapCalculationFunction
   @FunctionalInterface
   interface SingleMeasureCalculation {
     public abstract ScenarioArray<?> calculate(
-        ResolvedSwapTrade trade,
+        ResolvedIborFutureTrade trade,
         RatesScenarioMarketData marketData);
   }
