+++ modules/measure/src/main/java/com/opengamma/strata/measure/package-info.java
@@ -15,7 +15,7 @@
  *   {@code SwapMeasureCalculations}
  * <li>for a mixed portfolio of trades/positions and many scenarios,
  *   see {@link com.opengamma.strata.calc.CalculationRunner CalculationRunner} and
- *   functions such as {@link com.opengamma.strata.measure.swap.SwapCalculationFunction SwapCalculationFunction}
+ *   functions such as {@link com.opengamma.strata.measure.swap.SwapTradeCalculationFunction SwapCalculationFunction}
  * </ul>
  */
 package com.opengamma.strata.measure;
