+++ modules/measure/src/main/java/com/opengamma/strata/measure/swap/SwapTradeCalculationFunction.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.fx;
+package com.opengamma.strata.measure.swap;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -25,41 +25,55 @@ import com.opengamma.strata.data.scenario.ScenarioMarketData;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.measure.rate.RatesScenarioMarketData;
-import com.opengamma.strata.product.fx.FxNdf;
-import com.opengamma.strata.product.fx.FxNdfTrade;
-import com.opengamma.strata.product.fx.ResolvedFxNdfTrade;
+import com.opengamma.strata.product.swap.ResolvedSwapTrade;
+import com.opengamma.strata.product.swap.Swap;
+import com.opengamma.strata.product.swap.SwapTrade;
 
 /**
- * Perform calculations on a single {@code FxNdfTrade} for each of a set of scenarios.
+ * Perform calculations on a single {@code SwapTrade} for each of a set of scenarios.
  * <p>
  * This uses the standard discounting calculation method.
  * The supported built-in measures are:
  * <ul>
+ *   <li>{@linkplain Measures#PAR_RATE Par rate}
+ *   <li>{@linkplain Measures#PAR_SPREAD Par spread}
  *   <li>{@linkplain Measures#PRESENT_VALUE Present value}
  *   <li>{@linkplain Measures#PRESENT_VALUE_MULTI_CCY Present value with no currency conversion}
+ *   <li>{@linkplain Measures#EXPLAIN_PRESENT_VALUE Explain present value}
+ *   <li>{@linkplain Measures#CASH_FLOWS Cash flows}
  *   <li>{@linkplain Measures#PV01 PV01}
  *   <li>{@linkplain Measures#BUCKETED_PV01 Bucketed PV01}
+ *   <li>{@linkplain Measures#BUCKETED_GAMMA_PV01 Gamma PV01}
+ *   <li>{@linkplain Measures#ACCRUED_INTEREST Accrued interest}
+ *   <li>{@linkplain Measures#LEG_INITIAL_NOTIONAL Leg initial notional}
+ *   <li>{@linkplain Measures#LEG_PRESENT_VALUE Leg present value}
  *   <li>{@linkplain Measures#CURRENCY_EXPOSURE Currency exposure}
  *   <li>{@linkplain Measures#CURRENT_CASH Current cash}
- *   <li>{@linkplain Measures#FORWARD_FX_RATE Forward FX rate}
  * </ul>
  * <p>
- * The "natural" currency is the settlement currency of the trade.
+ * The "natural" currency is the currency of the swaption, which is limited to be single-currency.
  */
-public class FxNdfCalculationFunction
-    implements CalculationFunction<FxNdfTrade> {
+public class SwapTradeCalculationFunction
+    implements CalculationFunction<SwapTrade> {
 
   /**
    * The calculations by measure.
    */
   private static final ImmutableMap<Measure, SingleMeasureCalculation> CALCULATORS =
       ImmutableMap.<Measure, SingleMeasureCalculation>builder()
-          .put(Measures.PRESENT_VALUE, FxNdfMeasureCalculations::presentValue)
-          .put(Measures.PV01, FxNdfMeasureCalculations::pv01)
-          .put(Measures.BUCKETED_PV01, FxNdfMeasureCalculations::bucketedPv01)
-          .put(Measures.CURRENCY_EXPOSURE, FxNdfMeasureCalculations::currencyExposure)
-          .put(Measures.CURRENT_CASH, FxNdfMeasureCalculations::currentCash)
-          .put(Measures.FORWARD_FX_RATE, FxNdfMeasureCalculations::forwardFxRate)
+          .put(Measures.PAR_RATE, SwapMeasureCalculations::parRate)
+          .put(Measures.PAR_SPREAD, SwapMeasureCalculations::parSpread)
+          .put(Measures.PRESENT_VALUE, SwapMeasureCalculations::presentValue)
+          .put(Measures.EXPLAIN_PRESENT_VALUE, SwapMeasureCalculations::explainPresentValue)
+          .put(Measures.CASH_FLOWS, SwapMeasureCalculations::cashFlows)
+          .put(Measures.PV01, SwapMeasureCalculations::pv01)
+          .put(Measures.BUCKETED_PV01, SwapMeasureCalculations::bucketedPv01)
+          .put(Measures.BUCKETED_GAMMA_PV01, SwapMeasureCalculations::bucketedGammaPv01)
+          .put(Measures.ACCRUED_INTEREST, SwapMeasureCalculations::accruedInterest)
+          .put(Measures.LEG_INITIAL_NOTIONAL, SwapMeasureCalculations::legInitialNotional)
+          .put(Measures.LEG_PRESENT_VALUE, SwapMeasureCalculations::legPresentValue)
+          .put(Measures.CURRENCY_EXPOSURE, SwapMeasureCalculations::currencyExposure)
+          .put(Measures.CURRENT_CASH, SwapMeasureCalculations::currentCash)
           .build();
 
   private static final ImmutableSet<Measure> MEASURES = ImmutableSet.<Measure>builder()
@@ -70,13 +84,13 @@ public class FxNdfCalculationFunction
   /**
    * Creates an instance.
    */
-  public FxNdfCalculationFunction() {
+  public SwapTradeCalculationFunction() {
   }
 
   //-------------------------------------------------------------------------
   @Override
-  public Class<FxNdfTrade> targetType() {
-    return FxNdfTrade.class;
+  public Class<SwapTrade> targetType() {
+    return SwapTrade.class;
   }
 
   @Override
@@ -85,40 +99,38 @@ public class FxNdfCalculationFunction
   }
 
   @Override
-  public Currency naturalCurrency(FxNdfTrade trade, ReferenceData refData) {
-    return trade.getProduct().getSettlementCurrency();
+  public Currency naturalCurrency(SwapTrade trade, ReferenceData refData) {
+    return trade.getProduct().getLegs().get(0).getCurrency();
   }
 
   //-------------------------------------------------------------------------
   @Override
   public FunctionRequirements requirements(
-      FxNdfTrade trade,
+      SwapTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ReferenceData refData) {
 
     // extract data from product
-    FxNdf fx = trade.getProduct();
-    Currency settleCurrency = fx.getSettlementCurrency();
-    Currency otherCurrency = fx.getNonDeliverableCurrency();
-    ImmutableSet<Currency> currencies = ImmutableSet.of(settleCurrency, otherCurrency);
+    Swap product = trade.getProduct();
+    ImmutableSet<Currency> currencies = product.allPaymentCurrencies();
 
     // use lookup to build requirements
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
-    return ratesLookup.requirements(currencies);
+    return ratesLookup.requirements(currencies, product.allIndices());
   }
 
   //-------------------------------------------------------------------------
   @Override
   public Map<Measure, Result<?>> calculate(
-      FxNdfTrade trade,
+      SwapTrade trade,
       Set<Measure> measures,
       CalculationParameters parameters,
       ScenarioMarketData scenarioMarketData,
       ReferenceData refData) {
 
     // resolve the trade once for all measures and all scenarios
-    ResolvedFxNdfTrade resolved = trade.resolve(refData);
+    ResolvedSwapTrade resolved = trade.resolve(refData);
 
     // use lookup to query market data
     RatesMarketDataLookup ratesLookup = parameters.getParameter(RatesMarketDataLookup.class);
@@ -137,7 +149,7 @@ public class FxNdfCalculationFunction
   // calculate one measure
   private Result<?> calculate(
       Measure measure,
-      ResolvedFxNdfTrade trade,
+      ResolvedSwapTrade trade,
       RatesScenarioMarketData marketData) {
 
     SingleMeasureCalculation calculator = CALCULATORS.get(measure);
@@ -151,7 +163,7 @@ public class FxNdfCalculationFunction
   @FunctionalInterface
   interface SingleMeasureCalculation {
     public abstract ScenarioArray<?> calculate(
-        ResolvedFxNdfTrade trade,
+        ResolvedSwapTrade trade,
         RatesScenarioMarketData marketData);
   }
