+++ modules/measure/src/test/java/com/opengamma/strata/measure/curve/CurveEndToEndTest.java
@@ -57,7 +57,7 @@ import com.opengamma.strata.market.interpolator.CurveInterpolators;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.fra.FraTradeCalculationFunction;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
-import com.opengamma.strata.measure.swap.SwapCalculationFunction;
+import com.opengamma.strata.measure.swap.SwapTradeCalculationFunction;
 import com.opengamma.strata.product.Trade;
 import com.opengamma.strata.product.fra.FraTrade;
 import com.opengamma.strata.product.swap.SwapTrade;
@@ -172,7 +172,7 @@ public class CurveEndToEndTest {
   //-----------------------------------------------------------------------------------------------------------
   private static CalculationFunctions functions() {
     return CalculationFunctions.of(ImmutableMap.of(
-        SwapTrade.class, new SwapCalculationFunction(),
+        SwapTrade.class, new SwapTradeCalculationFunction(),
         FraTrade.class, new FraTradeCalculationFunction()));
   }
