+++ modules/measure/src/test/java/com/opengamma/strata/measure/fx/FxNdfTradeCalculationFunctionTest.java
@@ -8,6 +8,7 @@ package com.opengamma.strata.measure.fx;
 import static com.opengamma.strata.basics.currency.Currency.GBP;
 import static com.opengamma.strata.basics.currency.Currency.USD;
 import static com.opengamma.strata.basics.date.DayCounts.ACT_360;
+import static com.opengamma.strata.basics.index.FxIndices.GBP_USD_WM;
 import static com.opengamma.strata.collect.TestHelper.coverPrivateConstructor;
 import static com.opengamma.strata.collect.TestHelper.date;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -29,10 +30,10 @@ import com.opengamma.strata.calc.runner.CalculationParameters;
 import com.opengamma.strata.calc.runner.FunctionRequirements;
 import com.opengamma.strata.collect.result.Result;
 import com.opengamma.strata.data.FxRateId;
+import com.opengamma.strata.data.scenario.CurrencyValuesArray;
 import com.opengamma.strata.data.scenario.MultiCurrencyValuesArray;
 import com.opengamma.strata.data.scenario.ScenarioArray;
 import com.opengamma.strata.data.scenario.ScenarioMarketData;
-import com.opengamma.strata.data.scenario.ValuesArray;
 import com.opengamma.strata.market.curve.ConstantCurve;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveId;
@@ -42,27 +43,29 @@ import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.curve.TestMarketDataMap;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
-import com.opengamma.strata.pricer.fx.DiscountingFxSwapProductPricer;
+import com.opengamma.strata.pricer.fx.DiscountingFxNdfProductPricer;
 import com.opengamma.strata.pricer.rate.RatesProvider;
 import com.opengamma.strata.product.TradeInfo;
-import com.opengamma.strata.product.fx.FxSingle;
-import com.opengamma.strata.product.fx.FxSwap;
-import com.opengamma.strata.product.fx.FxSwapTrade;
-import com.opengamma.strata.product.fx.ResolvedFxSwap;
+import com.opengamma.strata.product.fx.FxNdf;
+import com.opengamma.strata.product.fx.FxNdfTrade;
+import com.opengamma.strata.product.fx.ResolvedFxNdf;
 
 /**
- * Test {@link FxSwapCalculationFunction}.
+ * Test {@link FxNdfTradeCalculationFunction}.
  */
 @Test
-public class FxSwapCalculationFunctionTest {
+public class FxNdfTradeCalculationFunctionTest {
 
   private static final ReferenceData REF_DATA = ReferenceData.standard();
-  private static final CurrencyAmount GBP_P1000 = CurrencyAmount.of(GBP, 1_000);
-  private static final CurrencyAmount USD_M1600 = CurrencyAmount.of(USD, -1_600);
-  private static final FxSingle LEG1 = FxSingle.of(GBP_P1000, USD_M1600, date(2015, 6, 30));
-  private static final FxSingle LEG2 = FxSingle.of(GBP_P1000.negated(), USD_M1600.negated(), date(2015, 9, 30));
-  private static final FxSwap PRODUCT = FxSwap.of(LEG1, LEG2);
-  public static final FxSwapTrade TRADE = FxSwapTrade.builder()
+  private static final FxRate FX_RATE = FxRate.of(GBP, USD, 1.5d);
+  private static final CurrencyAmount NOTIONAL = CurrencyAmount.of(GBP, (double) 100_000_000);
+  private static final FxNdf PRODUCT = FxNdf.builder()
+      .agreedFxRate(FX_RATE)
+      .settlementCurrencyNotional(NOTIONAL)
+      .index(GBP_USD_WM)
+      .paymentDate(date(2015, 3, 19))
+      .build();
+  public static final FxNdfTrade TRADE = FxNdfTrade.builder()
       .info(TradeInfo.builder()
           .tradeDate(date(2015, 6, 1))
           .build())
@@ -74,11 +77,11 @@ public class FxSwapCalculationFunctionTest {
       ImmutableMap.of(GBP, DISCOUNT_CURVE_GBP_ID, USD, DISCOUNT_CURVE_USD_ID),
       ImmutableMap.of());
   private static final CalculationParameters PARAMS = CalculationParameters.of(RATES_LOOKUP);
-  private static final LocalDate VAL_DATE = TRADE.getProduct().getNearLeg().getPaymentDate().minusDays(7);
+  private static final LocalDate VAL_DATE = TRADE.getProduct().getPaymentDate().minusDays(7);
 
   //-------------------------------------------------------------------------
   public void test_requirementsAndCurrency() {
-    FxSwapCalculationFunction function = new FxSwapCalculationFunction();
+    FxNdfTradeCalculationFunction function = new FxNdfTradeCalculationFunction();
     Set<Measure> measures = function.supportedMeasures();
     FunctionRequirements reqs = function.requirements(TRADE, measures, PARAMS, REF_DATA);
     assertThat(reqs.getOutputCurrencies()).containsExactly(GBP, USD);
@@ -89,42 +92,41 @@ public class FxSwapCalculationFunctionTest {
   }
 
   public void test_simpleMeasures() {
-    FxSwapCalculationFunction function = new FxSwapCalculationFunction();
+    FxNdfTradeCalculationFunction function = new FxNdfTradeCalculationFunction();
     ScenarioMarketData md = marketData();
     RatesProvider provider = RATES_LOOKUP.ratesProvider(md.scenario(0));
-    DiscountingFxSwapProductPricer pricer = DiscountingFxSwapProductPricer.DEFAULT;
-    ResolvedFxSwap resolved = TRADE.getProduct().resolve(REF_DATA);
-    MultiCurrencyAmount expectedPv = pricer.presentValue(resolved, provider);
-    double expectedParSpread = pricer.parSpread(resolved, provider);
+    DiscountingFxNdfProductPricer pricer = DiscountingFxNdfProductPricer.DEFAULT;
+    ResolvedFxNdf resolved = TRADE.getProduct().resolve(REF_DATA);
+    CurrencyAmount expectedPv = pricer.presentValue(resolved, provider);
     MultiCurrencyAmount expectedCurrencyExp = pricer.currencyExposure(resolved, provider);
-    MultiCurrencyAmount expectedCash = pricer.currentCash(resolved, provider.getValuationDate());
+    CurrencyAmount expectedCash = pricer.currentCash(resolved, provider);
+    FxRate expectedForwardFx = pricer.forwardFxRate(resolved, provider);
 
     Set<Measure> measures = ImmutableSet.of(
         Measures.PRESENT_VALUE,
         Measures.PRESENT_VALUE_MULTI_CCY,
-        Measures.PAR_SPREAD,
         Measures.CURRENCY_EXPOSURE,
         Measures.CURRENT_CASH,
         Measures.FORWARD_FX_RATE);
     assertThat(function.calculate(TRADE, measures, PARAMS, md, REF_DATA))
         .containsEntry(
-            Measures.PRESENT_VALUE, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedPv))))
+            Measures.PRESENT_VALUE, Result.success(CurrencyValuesArray.of(ImmutableList.of(expectedPv))))
         .containsEntry(
-            Measures.PRESENT_VALUE_MULTI_CCY, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedPv))))
-        .containsEntry(
-            Measures.PAR_SPREAD, Result.success(ValuesArray.of(ImmutableList.of(expectedParSpread))))
+            Measures.PRESENT_VALUE_MULTI_CCY, Result.success(CurrencyValuesArray.of(ImmutableList.of(expectedPv))))
         .containsEntry(
             Measures.CURRENCY_EXPOSURE, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedCurrencyExp))))
         .containsEntry(
-            Measures.CURRENT_CASH, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedCash))));
+            Measures.CURRENT_CASH, Result.success(CurrencyValuesArray.of(ImmutableList.of(expectedCash))))
+        .containsEntry(
+            Measures.FORWARD_FX_RATE, Result.success(ScenarioArray.of(ImmutableList.of(expectedForwardFx))));
   }
 
   public void test_pv01() {
-    FxSwapCalculationFunction function = new FxSwapCalculationFunction();
+    FxNdfTradeCalculationFunction function = new FxNdfTradeCalculationFunction();
     ScenarioMarketData md = marketData();
     RatesProvider provider = RATES_LOOKUP.ratesProvider(md.scenario(0));
-    DiscountingFxSwapProductPricer pricer = DiscountingFxSwapProductPricer.DEFAULT;
-    ResolvedFxSwap resolved = TRADE.getProduct().resolve(REF_DATA);
+    DiscountingFxNdfProductPricer pricer = DiscountingFxNdfProductPricer.DEFAULT;
+    ResolvedFxNdf resolved = TRADE.getProduct().resolve(REF_DATA);
     PointSensitivities pvPointSens = pricer.presentValueSensitivity(resolved, provider);
     CurrencyParameterSensitivities pvParamSens = provider.parameterSensitivity(pvPointSens);
     MultiCurrencyAmount expectedPv01 = pvParamSens.total().multipliedBy(1e-4);
@@ -154,7 +156,7 @@ public class FxSwapCalculationFunctionTest {
 
   //-------------------------------------------------------------------------
   public void coverage() {
-    coverPrivateConstructor(FxSwapMeasureCalculations.class);
+    coverPrivateConstructor(FxNdfMeasureCalculations.class);
   }
 
 }
