+++ modules/measure/src/test/java/com/opengamma/strata/measure/fx/FxSingleTradeCalculationFunctionTest.java
@@ -8,7 +8,6 @@ package com.opengamma.strata.measure.fx;
 import static com.opengamma.strata.basics.currency.Currency.GBP;
 import static com.opengamma.strata.basics.currency.Currency.USD;
 import static com.opengamma.strata.basics.date.DayCounts.ACT_360;
-import static com.opengamma.strata.basics.index.FxIndices.GBP_USD_WM;
 import static com.opengamma.strata.collect.TestHelper.coverPrivateConstructor;
 import static com.opengamma.strata.collect.TestHelper.date;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -30,10 +29,10 @@ import com.opengamma.strata.calc.runner.CalculationParameters;
 import com.opengamma.strata.calc.runner.FunctionRequirements;
 import com.opengamma.strata.collect.result.Result;
 import com.opengamma.strata.data.FxRateId;
-import com.opengamma.strata.data.scenario.CurrencyValuesArray;
 import com.opengamma.strata.data.scenario.MultiCurrencyValuesArray;
 import com.opengamma.strata.data.scenario.ScenarioArray;
 import com.opengamma.strata.data.scenario.ScenarioMarketData;
+import com.opengamma.strata.data.scenario.ValuesArray;
 import com.opengamma.strata.market.curve.ConstantCurve;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveId;
@@ -43,34 +42,23 @@ import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.curve.TestMarketDataMap;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
-import com.opengamma.strata.pricer.fx.DiscountingFxNdfProductPricer;
+import com.opengamma.strata.pricer.fx.DiscountingFxSingleProductPricer;
 import com.opengamma.strata.pricer.rate.RatesProvider;
-import com.opengamma.strata.product.TradeInfo;
-import com.opengamma.strata.product.fx.FxNdf;
-import com.opengamma.strata.product.fx.FxNdfTrade;
-import com.opengamma.strata.product.fx.ResolvedFxNdf;
+import com.opengamma.strata.product.fx.FxSingle;
+import com.opengamma.strata.product.fx.FxSingleTrade;
+import com.opengamma.strata.product.fx.ResolvedFxSingle;
 
 /**
- * Test {@link FxNdfCalculationFunction}.
+ * Test {@link FxSingleTradeCalculationFunction}.
  */
 @Test
-public class FxNdfCalculationFunctionTest {
+public class FxSingleTradeCalculationFunctionTest {
 
   private static final ReferenceData REF_DATA = ReferenceData.standard();
-  private static final FxRate FX_RATE = FxRate.of(GBP, USD, 1.5d);
-  private static final CurrencyAmount NOTIONAL = CurrencyAmount.of(GBP, (double) 100_000_000);
-  private static final FxNdf PRODUCT = FxNdf.builder()
-      .agreedFxRate(FX_RATE)
-      .settlementCurrencyNotional(NOTIONAL)
-      .index(GBP_USD_WM)
-      .paymentDate(date(2015, 3, 19))
-      .build();
-  public static final FxNdfTrade TRADE = FxNdfTrade.builder()
-      .info(TradeInfo.builder()
-          .tradeDate(date(2015, 6, 1))
-          .build())
-      .product(PRODUCT)
-      .build();
+  private static final CurrencyAmount GBP_P1000 = CurrencyAmount.of(GBP, 1_000);
+  private static final CurrencyAmount USD_M1600 = CurrencyAmount.of(USD, -1_600);
+  private static final FxSingle PRODUCT = FxSingle.of(GBP_P1000, USD_M1600, date(2015, 6, 30));
+  public static final FxSingleTrade TRADE = FxSingleTrade.builder().product(PRODUCT).build();
   private static final CurveId DISCOUNT_CURVE_GBP_ID = CurveId.of("Default", "Discount-GBP");
   private static final CurveId DISCOUNT_CURVE_USD_ID = CurveId.of("Default", "Discount-USD");
   private static final RatesMarketDataLookup RATES_LOOKUP = RatesMarketDataLookup.of(
@@ -81,7 +69,7 @@ public class FxNdfCalculationFunctionTest {
 
   //-------------------------------------------------------------------------
   public void test_requirementsAndCurrency() {
-    FxNdfCalculationFunction function = new FxNdfCalculationFunction();
+    FxSingleTradeCalculationFunction function = new FxSingleTradeCalculationFunction();
     Set<Measure> measures = function.supportedMeasures();
     FunctionRequirements reqs = function.requirements(TRADE, measures, PARAMS, REF_DATA);
     assertThat(reqs.getOutputCurrencies()).containsExactly(GBP, USD);
@@ -92,41 +80,45 @@ public class FxNdfCalculationFunctionTest {
   }
 
   public void test_simpleMeasures() {
-    FxNdfCalculationFunction function = new FxNdfCalculationFunction();
+    FxSingleTradeCalculationFunction function = new FxSingleTradeCalculationFunction();
     ScenarioMarketData md = marketData();
     RatesProvider provider = RATES_LOOKUP.ratesProvider(md.scenario(0));
-    DiscountingFxNdfProductPricer pricer = DiscountingFxNdfProductPricer.DEFAULT;
-    ResolvedFxNdf resolved = TRADE.getProduct().resolve(REF_DATA);
-    CurrencyAmount expectedPv = pricer.presentValue(resolved, provider);
+    DiscountingFxSingleProductPricer pricer = DiscountingFxSingleProductPricer.DEFAULT;
+    ResolvedFxSingle resolved = TRADE.getProduct().resolve(REF_DATA);
+    MultiCurrencyAmount expectedPv = pricer.presentValue(resolved, provider);
+    double expectedParSpread = pricer.parSpread(resolved, provider);
     MultiCurrencyAmount expectedCurrencyExp = pricer.currencyExposure(resolved, provider);
-    CurrencyAmount expectedCash = pricer.currentCash(resolved, provider);
+    MultiCurrencyAmount expectedCash = pricer.currentCash(resolved, provider.getValuationDate());
     FxRate expectedForwardFx = pricer.forwardFxRate(resolved, provider);
 
     Set<Measure> measures = ImmutableSet.of(
         Measures.PRESENT_VALUE,
         Measures.PRESENT_VALUE_MULTI_CCY,
+        Measures.PAR_SPREAD,
         Measures.CURRENCY_EXPOSURE,
         Measures.CURRENT_CASH,
         Measures.FORWARD_FX_RATE);
     assertThat(function.calculate(TRADE, measures, PARAMS, md, REF_DATA))
         .containsEntry(
-            Measures.PRESENT_VALUE, Result.success(CurrencyValuesArray.of(ImmutableList.of(expectedPv))))
+            Measures.PRESENT_VALUE, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedPv))))
         .containsEntry(
-            Measures.PRESENT_VALUE_MULTI_CCY, Result.success(CurrencyValuesArray.of(ImmutableList.of(expectedPv))))
+            Measures.PRESENT_VALUE_MULTI_CCY, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedPv))))
+        .containsEntry(
+            Measures.PAR_SPREAD, Result.success(ValuesArray.of(ImmutableList.of(expectedParSpread))))
         .containsEntry(
             Measures.CURRENCY_EXPOSURE, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedCurrencyExp))))
         .containsEntry(
-            Measures.CURRENT_CASH, Result.success(CurrencyValuesArray.of(ImmutableList.of(expectedCash))))
+            Measures.CURRENT_CASH, Result.success(MultiCurrencyValuesArray.of(ImmutableList.of(expectedCash))))
         .containsEntry(
             Measures.FORWARD_FX_RATE, Result.success(ScenarioArray.of(ImmutableList.of(expectedForwardFx))));
   }
 
   public void test_pv01() {
-    FxNdfCalculationFunction function = new FxNdfCalculationFunction();
+    FxSingleTradeCalculationFunction function = new FxSingleTradeCalculationFunction();
     ScenarioMarketData md = marketData();
     RatesProvider provider = RATES_LOOKUP.ratesProvider(md.scenario(0));
-    DiscountingFxNdfProductPricer pricer = DiscountingFxNdfProductPricer.DEFAULT;
-    ResolvedFxNdf resolved = TRADE.getProduct().resolve(REF_DATA);
+    DiscountingFxSingleProductPricer pricer = DiscountingFxSingleProductPricer.DEFAULT;
+    ResolvedFxSingle resolved = TRADE.getProduct().resolve(REF_DATA);
     PointSensitivities pvPointSens = pricer.presentValueSensitivity(resolved, provider);
     CurrencyParameterSensitivities pvParamSens = provider.parameterSensitivity(pvPointSens);
     MultiCurrencyAmount expectedPv01 = pvParamSens.total().multipliedBy(1e-4);
@@ -156,7 +148,7 @@ public class FxNdfCalculationFunctionTest {
 
   //-------------------------------------------------------------------------
   public void coverage() {
-    coverPrivateConstructor(FxNdfMeasureCalculations.class);
+    coverPrivateConstructor(FxSingleMeasureCalculations.class);
   }
 
 }
